chore(app): verify database schema on startup

+++ backend/src/main/resources/db/migration/V2__add_status_column.sql
@@ -1,1 +1,1 @@
-ALTER TABLE appointment ADD COLUMN status SMALLINT NOT NULL;
+ALTER TABLE appointment ADD COLUMN status INT NOT NULL;
